solve the puzzle of vue slots for the header, todo mixin for border on img
