+++ 05-Vue_Expanded/assignment/fishcamp/src/components/Nav.vue
@@ -1,0 +1,22 @@
+<script setup></script>
+
+<template>
+  <!-- nav bar part of header-->
+  <div class="row bg-warning p-0 m-0">
+    <nav class="navbar navbar-dark btn-group-horizontal justify-content-around">
+      <a href="#/"><button class="btn btn-outline-primary m-3">Landing Page</button></a>
+
+      <a href="#/Gallery"><button class="btn btn-outline-primary m-3">Trip Gallery</button></a>
+
+      <a href="#/Accordion"><button class="btn btn-outline-primary m-3">FAQ Page</button></a>
+
+      <a href="#/Contact"><button class="btn btn-outline-primary m-3">Contact Us</button></a>
+    </nav>
+  </div>
+</template>
+
+<script>
+export default {}
+</script>
+
+<style scoped></style>
